mcp/04-build-streammable-http-mcp-client/requirements.txt
# AI and MCP Core
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
mcp>=1.0.0
fastmcp>=0.9.0

# Data and Validation
pydantic>=2.5.0

# CLI and UI
click>=8.1.7
rich>=13.7.0

# Environment and Configuration
python-dotenv>=1.0.0

# Async Support
asyncio-mqtt>=0.13.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0

# adk
google-adk>=1.5.0